+++ dotnet/src/ComponentDiscovery.net48/My Project/AssemblyInfo.vb
@@ -1,4 +1,4 @@
-﻿Imports System
+Imports System
 Imports System.Reflection
 Imports System.Runtime.CompilerServices
 Imports System.Runtime.InteropServices
@@ -16,6 +16,6 @@ Imports System.Runtime.InteropServices
 <Assembly: ComVisible(False)>
 <Assembly: Guid("5027f65d-44ad-44b7-805a-2de61c83390e")>
 
-<Assembly: AssemblyVersion("4.9.0")>
-<Assembly: AssemblyInformationalVersion("4.9.0-localbuild")>
-'WARNING: DONT SPECIFY: <Assembly: AssemblyFileVersion(...)> 
+<Assembly: AssemblyVersion("4.10.3")>
+<Assembly: AssemblyInformationalVersion("4.10.3")>
+'WARNING: DONT SPECIFY: <Assembly: AssemblyFileVersion("4.10.3")>